requirements_gpu.txt
# GPU Processing Requirements for CAD Pipeline
# Core dependencies from original multithread module
numpy>=1.21.0
psutil>=5.8.0
pyvista>=0.40.0
trimesh>=3.15.0

# GPU acceleration frameworks (install based on your system)
# For NVIDIA GPUs with CUDA 11.8+ (choose one or both):
cupy-cuda11x>=12.0.0  # Preferred for NVIDIA RTX A6000
# OR cupy-cuda12x>=12.0.0  # For CUDA 12.x
# torch>=2.0.0,<3.0.0  # Alternative GPU acceleration

# NVIDIA GPU monitoring and management
nvidia-ml-py3>=11.4.1
pynvml>=11.4.1

# Additional GPU utilities
numba>=0.56.0  # JIT compilation for GPU kernels (optional)

# VTK with GPU support (may need special build)
vtk>=9.2.0

# Development and testing
pytest>=7.0.0
pytest-benchmark>=4.0.0
tqdm>=4.64.0

# For Runpod environment
matplotlib>=3.5.0
pillow>=9.0.0 